# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenFiltr API
  version: 1.0.0
  description: |
    REST API for OpenFiltr — a self-hosted DNS filtering platform.

    ## Authentication

    All protected endpoints require a bearer token in the `Authorization` header:

        Authorization: Bearer <token>

    Tokens are obtained from `POST /api/v1/auth/login` (JWT) or created at
    `POST /api/v1/auth/tokens` (long-lived API tokens prefixed with `oft_`).

    Browser sessions may also use the `openfiltr_token` HttpOnly cookie set on login.
  contact:
    name: OpenFiltr Community
    url: https://github.com/openfiltr/openfiltr
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpstreamServer:
      type: object
      required: [id, name, address, protocol, enabled]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
          example: "1.1.1.1:53"
        protocol:
          type: string
          enum: [udp, tcp, doh, dot]
          default: udp
        enabled:
          type: boolean
        priority:
          type: integer
          default: 0
        created_at:
          type: string
          format: date-time
    DNSEntry:
      type: object
      required: [id, host, entry_type, value, enabled]
      properties:
        id:
          type: string
          format: uuid
        host:
          type: string
          example: "router.home"
        entry_type:
          type: string
          enum: [A, AAAA, CNAME, MX, TXT, PTR, SRV]
        value:
          type: string
          example: "192.168.1.1"
        ttl:
          type: integer
          default: 300
        comment:
          type: string
          nullable: true
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ApiToken:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        scopes:
          type: string
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/auth/tokens:
    get:
      tags: [auth]
      summary: List API tokens
      description: Lists all API tokens for the current user. Raw token values are never returned after creation.
      operationId: listTokens
      responses:
        "200":
          description: Token list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiToken"
                  total:
                    type: integer
    post:
      tags: [auth]
      summary: Create API token
      description: Creates a new API token. The raw token value is returned **once only** — store it securely.
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "CI deployment"
                expires_at:
                  type: string
                  format: date-time
                  description: Optional expiry date (RFC3339)
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  token:
                    type: string
                    description: "Raw token value — shown only once. Begins with oft_"
        "400":
          description: Validation error
  /api/v1/auth/tokens/{id}:
    delete:
      tags: [auth]
      summary: Revoke API token
      operationId: deleteToken
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token revoked
        "404":
          description: Token not found

  # ── Filtering — Block Rules ────────────────────────────────────────────────
...
  /api/v1/dns/upstream-servers:
    get:
      tags: [dns]
      summary: List upstream DNS servers
      operationId: listUpstreamServers
      responses:
        "200":
          description: Upstream servers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/UpstreamServer"
                  total:
                    type: integer
    post:
      tags: [dns]
      summary: Add upstream server
      operationId: createUpstreamServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address]
              properties:
                name:
                  type: string
                address:
                  type: string
                  example: "1.1.1.1:53"
                protocol:
                  type: string
                  enum: [udp, tcp, doh, dot]
                  default: udp
                priority:
                  type: integer
                  default: 0
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamServer"
  /api/v1/dns/upstream-servers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [dns]
      summary: Get upstream server
      operationId: getUpstreamServer
      responses:
        "200":
          description: Upstream server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamServer"
    put:
      tags: [dns]
      summary: Update upstream server
      operationId: updateUpstreamServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                protocol:
                  type: string
                enabled:
                  type: boolean
                priority:
                  type: integer
      responses:
        "200":
          description: Updated
    delete:
      tags: [dns]
      summary: Delete upstream server
      operationId: deleteUpstreamServer
      responses:
        "200":
          description: Deleted

  # ── DNS — Entries ──────────────────────────────────────────────────────────
  /api/v1/dns/entries:
    get:
      tags: [dns]
      summary: List local DNS entries
      operationId: listDNSEntries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: DNS entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DNSEntry"
                  total:
                    type: integer
    post:
      tags: [dns]
      summary: Create local DNS entry
      operationId: createDNSEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host, entry_type, value]
              properties:
                host:
                  type: string
                  example: "router.home"
                entry_type:
                  type: string
                  enum: [A, AAAA, CNAME, MX, TXT, PTR, SRV]
                value:
                  type: string
                ttl:
                  type: integer
                  default: 300
                comment:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DNSEntry"
  /api/v1/dns/entries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [dns]
      summary: Get DNS entry
      operationId: getDNSEntry
      responses:
        "200":
          description: DNS entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DNSEntry"
    put:
      tags: [dns]
      summary: Update DNS entry
      operationId: updateDNSEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host:
                  type: string
                value:
                  type: string
                ttl:
                  type: integer
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated
    delete:
      tags: [dns]
      summary: Delete DNS entry
      operationId: deleteDNSEntry
      responses:
        "200":
          description: Deleted

  # ── Clients ────────────────────────────────────────────────────────────────
...